# docs/swagger.yaml
basePath: /
definitions:
  entity.AudioFile:
    properties:
      filename:
        type: string
      id:
        type: integer
      status:
        type: string
      user_id:
        type: string
    type: object
  entity.AudioSegment:
    properties:
      audio_id:
        type: integer
      audio_name:
        type: string
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: integer
      status:
        type: string
    type: object
  entity.AudioSegmentList:
    properties:
      audio_segments:
        items:
          $ref: '#/definitions/entity.AudioSegment'
        type: array
      count:
        type: integer
    type: object
  entity.DatasetViewerList:
    properties:
      audio_id:
        type: integer
      audio_url:
        type: string
      chunk_id:
        type: integer
      chunk_url:
        type: string
      duration:
        type: number
      next_text:
        type: string
      previous_text:
        type: string
      sentence:
        type: string
      text:
        type: string
      transcriber:
        type: string
    type: object
  entity.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  entity.Statistics:
    properties:
      duration:
        additionalProperties:
          type: integer
        type: object
      next_text:
        additionalProperties:
          type: integer
        type: object
      previous_text:
        additionalProperties:
          type: integer
        type: object
      text:
        additionalProperties:
          type: integer
        type: object
      transcriber:
        additionalProperties:
          type: integer
        type: object
    type: object
  entity.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  entity.TranscriptPersent:
    properties:
      audio_file_id:
        type: integer
      completed_segments:
        type: integer
      filename:
        type: string
      percent:
        type: number
      total_segments:
        type: integer
    type: object
...
  entity.UserTranscriptStatictics:
    properties:
      daily_chunks:
        type: string
      total_audio_files:
        type: integer
      total_chunks:
        type: integer
      total_minutes:
        type: number
      username:
        type: string
      weekly_audio_files:
        type: integer
      weekly_chunks:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server Voice Transcribe server.
  title: Voice Transcribe API
  version: "1.0"
paths:
  /api/v1/audio_file/{id}:
    get:
      consumes:
      - application/json
      description: Get audio file
      parameters:
      - description: Audio ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AudioFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audio file
      tags:
      - audio
  /api/v1/audio_segment:
    get:
      consumes:
      - application/json
      description: Get a list of audio_segment
      parameters:
      - description: Filter by audio id
        in: query
        name: audio_id
        type: integer
      - description: user id
        in: query
        name: user_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AudioSegmentList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of audio_segment
      tags:
      - audio_segment
  /api/v1/audio_segment/{id}:
    get:
      consumes:
      - application/json
      description: Get a audio_segment by ID
      parameters:
      - description: AudioSegment ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AudioSegment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a audio_segment by ID
      tags:
      - audio_segment
  /api/v1/audio_segment/delete:
    delete:
      consumes:
      - application/json
      description: Delete a audio_segment
      parameters:
      - description: AudioSegment ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a audio_segment
      tags:
      - audio_segment
...
  /api/v1/dashboard:
    get:
      consumes:
      - application/json
      description: Get a list of audio percent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.TranscriptPersent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of audio percent
      tags:
      - dashboard
  /api/v1/dashboard/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get the user dashboard
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserTranscriptStatictics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the user dashboard
      tags:
      - dashboard
  /api/v1/dataset_viewer:
    get:
      consumes:
      - application/json
      description: Get a list of dataset_viewer
      parameters:
      - description: Offset for pagination
        in: query
        name: offset
        type: number
      - description: Limit for pagination
        in: query
        name: limit
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DatasetViewerList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of dataset_viewer
      tags:
      - dashboard
  /api/v1/statistic:
    get:
      consumes:
      - application/json
      description: Get statistic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Statistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get statistic
      tags:
      - dashboard
...
  /api/v1/upload-zip-audio:
    post:
      consumes:
      - multipart/form-data
      description: Upload Zip file
      parameters:
      - description: Zip file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload Zip file
      tags:
      - audio
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"